suppress warnings

+++ src/2D/renderer2D.cpp
@@ -31,9 +31,9 @@ namespace GJGO
 
             glm::mat4 toReturn = glm::translate(glm::mat4(1.0f), glm::vec3(a_position.x, a_position.y, 0.0f));
 
-            toReturn = glm::translate(toReturn, glm::vec3(a_size.width / 2.0f, a_size.height / 2.0f, 0.0f));
+            toReturn = glm::translate(toReturn, glm::vec3(static_cast<float>(a_size.width) / 2.0f, static_cast<float>(a_size.height) / 2.0f, 0.0f));
             toReturn = glm::rotate(toReturn, glm::radians(a_rotation * -1.0f), glm::vec3(0.0f, 0.0f, 1.0f));
-            toReturn = glm::translate(toReturn, glm::vec3((a_size.width / 2.0f) * -1.0f, (a_size.height / 2.0f) * -1.0f, 0.0f));
+            toReturn = glm::translate(toReturn, glm::vec3((static_cast<float>(a_size.width) / 2.0f) * -1.0f, (static_cast<float>(a_size.height) / 2.0f) * -1.0f, 0.0f));
 
             return glm::scale(toReturn, glm::vec3(static_cast<float>((1 + a_size.width)) / 1.0f, static_cast<float>((1 + a_size.height)) / 1.0f, 1.0f));
         }
